change function placement; line wrapping
