+++ autocompleter/utils.py
@@ -185,45 +185,3 @@ def term_phrase_already_aliased(aliasable_phrase_start, aliasable_phrase_end,
         elif aliasable_phrase_end <= aliased_phrase_end:
             return True
     return False
-
-
-def validate_facets(facets):
-    """
-    Validates the facets list has all the keys we expect as well
-    as the correct facet types.
-    """
-    for facet in facets:
-        try:
-            facet_type = facet['type']
-            if facet_type not in ['and', 'or']:
-                return False
-            sub_facets = facet['facets']
-            if len(sub_facets) == 0:
-                return False
-            for sub_facet in sub_facets:
-                sub_facet['key']
-                sub_facet['value']
-        except (KeyError, TypeError):
-            return False
-    return True
-
-
-def hash_facets(facets):
-    """
-    Given an array of facet data, return a deterministic hash such that
-    the ordering of keys inside the facet dicts does not matter.
-    """
-    facet_hashes = []
-    for facet in facets:
-        sub_facet_hashes = []
-        facet_type = facet['type']
-        sub_facets = facet['facets']
-        for sub_facet in sub_facets:
-            sub_facet_str = 'key:' + sub_facet['key'] + 'value:' + sub_facet['value']
-            sub_facet_hashes.append(hash(sub_facet_str))
-        sub_facet_hashes.sort()
-        facet_str = 'type:' + facet_type + 'facets:' + str(sub_facet_hashes)
-        facet_hashes.append(hash(facet_str))
-    facet_hashes.sort()
-    final_facet_hash = hash(str(facet_hashes))
-    return final_facet_hash
