+++ setup.py
@@ -4,8 +4,8 @@ setup(
     name='django-autocompleter',
     version="0.6.0",
     description='A redis-backed autocompletor for Django projects',
-    author='Tom Jakeway',
-    author_email='tom@ycharts.com',
+    author='Ara Anjargolian',
+    author_email='ara818@gmail.com',
     url='http://github.com/ara818/django-autocompleter',
     packages=['autocompleter', 'autocompleter.management', 'autocompleter.management.commands'],
     install_requires=['setuptools', 'redis'],
